refactor(backend): consolidate 4 competing terminal API implementations (#3383)
